+++ changelog.md
@@ -1,3 +1,10 @@
+
+Version 0.3:
++Refactored Result Classes again with an abstract class
++Battleresult now has a list of phases
++Main class that has hardcoded attacker and defender
++Printer class for console output
+
 Version 0.2:
 + Javadoc for classes and methods
 + Changed access modifiers for several methods in CombatRules